feat(task): deleting inexistent task now is idempotent

+++ src/main/java/com/example/demo/service/TaskService.java
@@ -36,6 +36,8 @@ public class TaskService {
     }
 
     public void deleteTask(int id) {
-        taskRepository.deleteById(id);
+        if (taskRepository.findById(id).isPresent()) {
+            taskRepository.deleteById(id);
+        }
     }
 }
